# gp-integration-utils/microtest/NHSApp-SwaggerContract.yaml
swagger: '2.0'
info:
  version: 20190225
  title: NHS App API
  description: This is the API for the NHS App. - updated version supplied 15/02/2019.
schemes:
  - https
paths:
...
  /v1/patient/appointments:
    delete:
      tags:
        - appointments
      summary: Cancels an upcoming appointment for the user in context.
      description: >-
        Cancels a specified appointment for the user in context.
      consumes:
        - application/json
      parameters:
        - in: body
          name: AppointmentCancelRequest
          required: true
          description: >-
            Details of the appointment to be cancelled, with an optional cancellation reason.
          schema:
            type: object
            required:
              - appointmentId
            properties:
              appointmentId:
                type: string
                description: Identifier of the appointment to be cancelled.
                minLength: 1
              cancelReason:
                type: string
                description: Freetext reason the appointment is being cancelled
                minLength: 1
                maxLength: 200
      responses:
        '204':
          description: Successfully cancelled the specified appointment.
        '400':
          description: Bad request. Supplied request is incomplete or invalid.
        '403':
          description: >-
            The patient does not have the necessary permissions within the GP
            system.
        '409':
          description: >-
            The specified appointment does not exist or is not in a suitable
            state to be cancelled.
        '500':
          description: An unexpected error occurred during the processing of the request.
    get:
      tags:
        - appointments
      summary: Gets a list of existing appointments for the user in context.
      description: >-
        Returns list of booked appointments for the user in context, and
        associated metadata that can be used to cancel upcoming appointments.
      produces:
        - application/json
      parameters:
        - in: query
          name: pastAppointmentsFromDate
          type: string
          format: date-time
          required: false
          description: >-
            Earliest date for which to return historical appointments,
            defaults to 'today' if not set
      responses:
        '200':
          description: >-
            Returns list of booked appointments and associated meta data.
            Returns empty collection of appointments if GP system does not
            return a list of appointmetns.
          schema:
            type: object
            properties:
              appointments:
                type: array
                description: Array of booked Appointments.
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Appointment Identifier.
                    type:
                      type: string
                      description: Type of appointment slot, e.g. "Baby Clinics".
                    startTime:
                      type: string
                      format: date-time
                      description: Appointment Start Date Time.
                    duration:
                      type: string
                      description: Duration of appointment in minutes
                    endTime:
                      type: string
                      format: date-time
                      description: Appointment End Date Time.
                    location:
                      type: string
                      description: Physical location of the appointment.
                    clinicians:
                      type: array
                      description: 'Array of Clinicians, accounts for joint sessions.'
                      items:
                        type: string
        '400':
          description: Bad request. A querystring parameter is not in the expected format.
        '403':
          description: >-
            The patient does not have the necessary permissions within the GP
            system.
        '500':
          description: An unexpected error occurred processing the request.
    post:
      tags:
        - appointments
      summary: Creates a GP Appointment for the user in context.
      description: >-
        Books an appointment slot for the user in context, given a slot
        identifier. Identifiers of available slots can be obtained using the
        appointment-slots endpoint.
      consumes:
        - application/json
      parameters:
        - in: body
          name: AppointmentBookRequest
          required: true
          description: >-
            Details of the appointment slot to be booked, along with reasons for
            the booking being made.
          schema:
            type: object
            required:
              - slotId
              - bookingReason
            properties:
              slotId:
                type: string
                description: Identifier of appointment slot to be booked.
                minLength: 1
              bookingReason:
                type: string
                description: >-
                  Reason for booking appointment. e.g. symptoms being
                  experienced by the patient.
                minLength: 1
                maxLength: 150
              telephoneNumber:
                type: string
                description: Patient's phone number for telephone appointment.
                maxLength: 15

# Removed as not required, the slotId already knows this.
              # startTime:
              #   type: string
              #   format: date-time
              #   description: Appointment Start Date Time.
              # endTime:
              #   type: string
              #   format: date-time
              #   description: Appointment End Date Time.
      responses:
        '201':
          description: Appointment successfully created.
        '400':
          description: Bad request. Supplied request is incomplete or invalid.
        '403':
          description: >-
            The patient does not have the necessary permissions within the GP
            system.
        '409':
          description: Conflict. The chosen appointment slot is not available for booking.
        '500':
          description: An unexpected error occurred processing the request.
...